+++ mcio_remote/controller.py
@@ -48,7 +48,7 @@ class ControllerSync(ControllerCommon):
     def recv_observation(
         self, block: bool = True, timeout: float | None = None
     ) -> network.ObservationPacket:
-        """Receive observation. Blocks"""
+        """Receive observation. Always blocks - ignores block and timeout args"""
         obs = self._mcio_conn.recv_observation(block=True)
         if obs is None:
             # This will only ever happen when zmq is shutting down